requirements.txt
yt-dlp~=2025.8.27
fastapi~=0.116.1
uvicorn[standard]~=0.35.0
boto3~=1.40.21
botocore~=1.40.21
python-dotenv~=1.1.1
pymongo~=4.14.1
pydantic~=2.12.0a1
beanie~=2.0.0